merge: merge of all commits until 2/1/2021

+++ src/main/java/io/github/syst3ms/skriptparser/expressions/ExprAmount.java
@@ -4,12 +4,13 @@ import io.github.syst3ms.skriptparser.Parser;
 import io.github.syst3ms.skriptparser.lang.Expression;
 import io.github.syst3ms.skriptparser.lang.TriggerContext;
 import io.github.syst3ms.skriptparser.lang.Variable;
-import io.github.syst3ms.skriptparser.lang.base.PropertyExpression;
+import io.github.syst3ms.skriptparser.lang.properties.PropertyExpression;
 import io.github.syst3ms.skriptparser.log.ErrorType;
 import io.github.syst3ms.skriptparser.parsing.ParseContext;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
+import java.math.BigInteger;
 import java.util.Map;
 import java.util.Optional;
 
@@ -57,9 +58,11 @@ public class ExprAmount extends PropertyExpression<Number, Object> {
 		if (recursive) {
 			Optional<Object> var = ((Variable<?>) getOwner()).getRaw(ctx);
 			if (var.isPresent())
-				return new Number[] {getRecursiveSize((Map<String, ?>) var.get())}; // Should already be a Map
+				return new Number[] {
+						BigInteger.valueOf(getRecursiveSize((Map<String, ?>) var.get()))
+			};
 		}
-		return new Number[] {getOwner().getValues(ctx).length};
+		return new Number[] {BigInteger.valueOf(getOwner().getValues(ctx).length)};
 	}
 
 	@Override
